Updated lightnings to be less blocking
I wanted to use this pattern as part of a larger program where I switch between patterns with a button, and as such needed it to be less blocking(there still are some small delays). I rewrote it so other things can happen in de loop (such as checking for buttons). Thanks for the effect!